Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -4,7 +4,7 @@
 struct Vehicle{
     char number[20];
     char type[10];
-    int toll;
+    float toll;
 };
 
 int main(){
